# docs/swagger.yaml
openapi: 3.0.0
info:
  title: API de Superhéroes y Villanos
  version: 1.0.0
  description: Una API simple para gestionar superhéroes, villanos y sus aventuras.

servers:
  - url: http://localhost:3001/api # La URL base de tu API, recuerda tu app.use('/api', ...)

tags:
  - name: Héroes
    description: Operaciones relacionadas con los superhéroes
  - name: Villanos
    description: Operaciones relacionadas con los villanos

paths:
  /heroes:
    get:
      summary: Obtiene todos los héroes
      tags:
        - Héroes
      responses:
        '200':
          description: Lista de héroes obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Hero'
        '500':
          description: Error interno del servidor
    post:
      summary: Agrega un nuevo héroe
      tags:
        - Héroes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeroInput'
      responses:
        '201':
          description: Héroe agregado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hero'
        '400':
          description: Datos de entrada inválidos
        '500':
          description: Error interno del servidor

  /heroes/{id}: # <--- Bloque unificado para GET, PUT, DELETE
    get:
      summary: Obtiene un héroe por ID
      tags:
        - Héroes
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID del héroe a obtener
      responses:
        '200':
          description: Héroe encontrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hero'
        '404':
          description: Héroe no encontrado
        '500':
          description: Error interno del servidor
    put:
      summary: Actualiza un héroe existente
      tags:
        - Héroes
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID del héroe a actualizar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/HeroInput'
      responses:
        '200':
          description: Héroe actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Hero'
        '404':
          description: Héroe no encontrado
        '500':
          description: Error interno del servidor
    delete:
      summary: Elimina un héroe
      tags:
        - Héroes
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID del héroe a eliminar
      responses:
        '200':
          description: Héroe eliminado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Héroe eliminado con éxito
        '404':
          description: Héroe no encontrado
        '500':
          description: Error interno del servidor

  /heroes/city/{city}:
    get:
      summary: Busca héroes por ciudad
      tags:
        - Héroes
      parameters:
        - in: path
          name: city
          schema:
            type: string
          required: true
          description: Nombre de la ciudad para buscar héroes
      responses:
        '200':
          description: Lista de héroes encontrada exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Hero'
        '500':
          description: Error interno del servidor

  /heroes/{id}/enfrentar:
    post:
      summary: Hace que un héroe se enfrente a un villano
      tags:
        - Héroes
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID del héroe que enfrentará
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                villainId:
                  type: integer
                  description: ID del villano al que se enfrentará el héroe
                  example: 1
              required:
                - villainId
      responses:
        '200':
          description: Enfrentamiento registrado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  hero:
                    type: object
                    properties:
                      id: { type: integer }
                      alias: { type: string }
                      name: { type: string }
                      power: { type: integer }
                  villain:
                    type: object
                    properties:
                      id: { type: integer }
                      alias: { type: string }
                      name: { type: string }
                      power: { type: integer }
                  outcome: { type: string, example: 'Batman (Bruce Wayne) derrota a Joker (Clown Prince of Crime)! ¡La justicia prevalece!' }
                  winner: { type: string, example: 'Batman' }
                  timestamp: { type: string, format: 'date-time' }
        '404':
          description: Héroe o villano no encontrado
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: Héroe no encontrado
        '500':
          description: Error interno del servidor

  /heroes/team-battle: # <--- Ruta para batallas de equipo aqui esta
    post:
      summary: Simula una batalla entre un equipo de 3 héroes y 3 villanos.
      tags:
        - Héroes
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                heroIds:
                  type: array
                  items:
                    type: integer
                  minItems: 3
                  maxItems: 3
                  description: Array de 3 IDs de héroes para el equipo.
                  example: [1, 2, 3]
                villainIds:
                  type: array
                  items:  
                    type: integer
                  minItems: 3
                  maxItems: 3
                  description: Array de 3 IDs de villanos para el equipo.
                  example: [1, 2, 3]
                   # INICIO DE CAMBIO: NUEVO CAMPO damageType aqui se agrega un cambio revisarlo
                damageType: # <-- Nuevo campo para el tipo de daño
                  type: string
                  description: Tipo de daño a aplicar en cada ataque.
                  enum: [basic, power, critical] # Valores permitidos
                  example: basic
                  
              required:
                - heroIds
                - villainIds
                - damageType # <-- este tambien es un cambio que realize
      responses:
        '200':
          description: Resultado de la batalla de equipo.
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: "El equipo de Héroes ha ganado la batalla!"
                  battleDetails:
                    type: object
                    properties:
                      heroes:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: integer }
                            alias: { type: string }
                            name: { type: string }
                            power: { type: integer }
                            initialHealth: { type: integer }
                            remainingHealth: { type: integer }
                            criticalDamageChance: { type: string }
                      villains:
                        type: array
                        items:
                          type: object
                          properties:
                            id: { type: integer }
                            alias: { type: string }
                            name: { type: string }
                            power: { type: integer }
                            initialHealth: { type: integer }
                            remainingHealth: { type: integer }
                            criticalDamageChance: { type: string }
                      winner:
                        type: string
                        example: "Equipo de Héroes"
                      loser:
                        type: string
                        example: "Equipo de Villanos"
                      roundsFought:
                        type: integer
                        example: 5
                      log:
                        type: array
                        items:
                          type: string
                          example: "Ronda 1: Superman ataca a Joker por 15 de daño. Joker tiene 85 de vida restante."
        '400':
          description: Solicitud inválida (ej. número incorrecto de participantes).
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "Debe proporcionar exactamente 3 IDs para héroes y 3 para villanos."
        '404':
          description: Héroes o villanos no encontrados.
          content:
            application/json:
              schema:
                type: object
                properties:
                  error:
                    type: string
                    example: "No se encontraron algunos héroes/villanos con los IDs proporcionados."
        '500':
          description: Error interno del servidor.

  /villains:
    get:
      summary: Obtiene todos los villanos
      tags:
        - Villanos
      responses:
        '200':
          description: Lista de villanos obtenida exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Villain'
        '500':
          description: Error interno del servidor
    post:
      summary: Agrega un nuevo villano
      tags:
        - Villanos
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VillainInput'
      responses:
        '201':
          description: Villano agregado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Villain'
        '400':
          description: Datos de entrada inválidos
        '500':
          description: Error interno del servidor

  /villains/{id}:
    get:
      summary: Obtiene un villano por ID
      tags:
        - Villanos
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID del villano a obtener
      responses:
        '200':
          description: Villano encontrado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Villain'
        '404':
          description: Villano no encontrado
        '500':
          description: Error interno del servidor
    put:
      summary: Actualiza un villano existente
      tags:
        - Villanos
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID del villano a actualizar
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VillainInput'
      responses:
        '200':
          description: Villano actualizado exitosamente
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Villain'
        '404':
          description: Villano no encontrado
        '500':
          description: Error interno del servidor
    delete:
      summary: Elimina un villano
      tags:
        - Villanos
      parameters:
        - in: path
          name: id
          schema:
            type: integer
          required: true
          description: ID del villano a eliminar
      responses:
        '200':
          description: Villano eliminado exitosamente
          content:
            application/json:
              schema:
                type: object
                properties:
                  message:
                    type: string
                    example: Villano eliminado con éxito
        '404':
          description: Villano no encontrado
        '500':
          description: Error interno del servidor

  /villains/city/{city}:
    get:
      summary: Busca villanos por ciudad
      tags:
        - Villanos
      parameters:
        - in: path
          name: city
          schema:
            type: string
          required: true
          description: Nombre de la ciudad para buscar villanos
      responses:
        '200':
          description: Lista de villanos encontrada exitosamente
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Villain'
        '500':
          description: Error interno del servidor


components:
  schemas:
    Hero:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          example: Peter Parker
        alias:
          type: string
          example: Spider-Man
        city:
          type: string
          example: New York
        team:
          type: string
          example: Avengers
        # powerLevel: # Descomenta si decides usar powerLevel más adelante
        #   type: integer
        #   example: 50
      required:
        - name
        - alias
    HeroInput:
      type: object
      properties:
        name:
          type: string
          example: Peter Parker
        alias:
          type: string
          example: Spider-Man
        city:
          type: string
          example: New York
        team:
          type: string
          example: Avengers
        # powerLevel: # Descomenta si decides usar powerLevel más adelante
        #   type: integer
        #   example: 50
      required:
        - name
        - alias
    Villain:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        name:
          type: string
          example: Joker
        alias:
          type: string
          example: Clown Prince of Crime
        powers:
          type: array
          items:
            type: string
          example: ["Chaos Incarnate", "Criminal Mastermind"]
        # powerLevel: # Descomenta si decides usar powerLevel más adelante
        #   type: integer
        #   example: 45
      required:
        - name
        - alias
    VillainInput:
      type: object
      properties:
        name:
          type: string
          example: Joker
        alias:
          type: string
          example: Clown Prince of Crime
        powers:
          type: array
          items:
            type: string
          example: ["Chaos Incarnate", "Criminal Mastermind"]
        # powerLevel: # Descomenta si decides usar powerLevel más adelante
        #   type: integer
        #   example: 45
      required:
        - name
        - alias