# models/villainModel.js
// models/villainModel.js
class Villain {
  constructor(id, name, alias, city, team) {
    this.id = id
        this.name = name
        this.alias = alias
        this.city = city
        this.team = team
  }
}


export default Villain